Int should be double for locations

+++ src/main/kotlin/be/ugent/webdevelopment/backend/geocode/controllers/wrappers/CurrentLocationReceiveWrapper.kt
@@ -1,6 +1,6 @@
 package be.ugent.webdevelopment.backend.geocode.controllers.wrappers
 
 class CurrentLocationReceiveWrapper(
-        var lat: Int,
-        var lon: Int
+        var lat: Double,
+        var lon: Double
 ): Wrapper()
